+++ pkg/bench/append.go
@@ -105,7 +105,7 @@ func (u *Append) Start(ctx context.Context, wait chan struct{}) error {
 				client, cldone := u.Client()
 				op := Operation{
 					OpType:   "APPEND",
-					Thread:   uint16(i),
+					Thread:   uint32(i),
 					Size:     obj.Size,
 					ObjPerOp: 1,
 					File:     obj.Name,
