+++ pkg/bench/retention.go
@@ -91,7 +91,7 @@ func (g *Retention) Prepare(ctx context.Context) error {
 					client, cldone := g.Client()
 					op := Operation{
 						OpType:   http.MethodPut,
-						Thread:   uint16(i),
+						Thread:   uint32(i),
 						Size:     obj.Size,
 						File:     obj.Name,
 						ObjPerOp: 1,
@@ -176,7 +176,7 @@ func (g *Retention) Start(ctx context.Context, wait chan struct{}) error {
 				client, cldone := g.Client()
 				op := Operation{
 					OpType:   "RETENTION",
-					Thread:   uint16(i),
+					Thread:   uint32(i),
 					Size:     0,
 					File:     obj.Name,
 					ObjPerOp: 1,
